Outline restructuring
Went through this with @mohitbhoite and did some re-structuring. I think that we'll need more time on set-up because there are a lot of basics to cover. First thing tomorrow morning we should go through and prioritize; I think we might have more here than we can do in a day!

I also think that, once we cover the basics, each of the examples can be very brief; 5 minutes max.

Let me know what you think!